Edit decks (with minor auth changes) WIP
