+++ app/javascript/pages/Deck.jsx
@@ -1,17 +1,17 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { Link, useParams } from "react-router-dom";
-import ReactTooltip from "react-tooltip";
 import _ from "lodash";
 
 import makeApiRequest from "../api/makeApiRequest";
+import { AuthContext } from "../components/AuthProvider";
 
-import CardPanel from "../components/CardPanel";
+import DeckCardList from "../components/DeckCardList";
 
 // List all of the cards in a deck.
-// SO similar to Set.  TODO - Extract the common bits
 const Deck = () => {
   const [ deckData, setDeckData ] = useState({})
   const params = useParams();
+  const { user } = useContext(AuthContext);
 
   useEffect(() => {
     async function getDeckById() {
@@ -22,152 +22,30 @@ const Deck = () => {
     getDeckById();
   }, [])
 
-  const deckCards = deckData.cards || []
-  const cardsByBlock = _.groupBy(deckCards, 'block')
-
-  const getQuantityOfObjective = (objectiveCard) => {
-    const foundBlock = deckData.card_blocks.find((blk) => blk.block == objectiveCard.block)
-
-    return foundBlock ? foundBlock.quantity : 0;
-  };
+  const isUserDeck = () => deckData?.user_id === user?.id;
 
   return (
     <div className="container">
-      <h2>{ deckData.name }</h2>
-      <h4>Objectives</h4>
-      <table className="table table-sm">
-        <thead>
-          <tr>
-            <th>Quantity</th>
-            <th>Name</th>
-            <th>{"  "}</th>
-          </tr>
-        </thead>
-        <tbody>
-          { _.map(cardsByBlock, (cards, blockNum) => {
-              const objectiveCard = cards.filter((c) => c.card_type === "Objective")[0];
-              const nonObjectiveCards = cards
-                                        .filter((c) => c.card_type !== "Objective")
-                                        .sort((a, b) => a.block_number - b.block_number);
-
-              return (
-                <>
-                  <tr>
-                    <td>
-                      {getQuantityOfObjective(objectiveCard)}
-                    </td>
-                    <td>
-                      <Link
-                        className="link-primary"
-                        data-tip
-                        data-for={`card-${objectiveCard.id}`}
-                        to={`/cards/${objectiveCard.id}`}
-                      >
-                        {objectiveCard.name}
-                      </Link>
-                      <ReactTooltip
-                        backgroundColor="white"
-                        border
-                        borderColor="black"
-                        className="card-tooltip"
-                        textColor="black"
-                        id={`card-${objectiveCard.id}`}
-                        place="right"
-                      >
-                        <CardPanel cardData={ objectiveCard } />
-                      </ReactTooltip>
-                    </td>
-                    <td>
-                      <i
-                        className="bi bi-chevron-right collapsed"
-                        data-bs-toggle="collapse"
-                        data-bs-target={`#collapse-${blockNum}`}
-                        aria-expanded="false"
-                        aria-controls={`collapse-${blockNum}`}>
-                      </i>
-                    </td>
-                  </tr>
-                  {
-                    nonObjectiveCards.map((card) =>
-                      <tr className="collapse" id={`collapse-${blockNum}`} key={`card-row-${card.id}`}>
-                        <td>
-                        </td>
-                        <td>&nbsp;&nbsp;<Link
-                          className="link-primary"
-                          data-tip
-                          data-for={`card-${card.id}`}
-                          to={`/cards/${card.id}`}
-                          >
-                            {card.name}
-                          </Link>
-                          <ReactTooltip
-                            backgroundColor="white"
-                            border
-                            borderColor="black"
-                            className="card-tooltip"
-                            textColor="black"
-                            id={`card-${card.id}`}
-                            place="right"
-                          >
-                            <CardPanel cardData={ card } />
-                          </ReactTooltip>
-                        </td>
-                      </tr>
-                    )
-                  }
-                </>
-              )
-            })
-          }
-        </tbody>
-      </table>
-
-
-      <h4>Cards</h4>
-      <table className="table table-sm">
-        <thead>
-          <tr>
-            <th>Quantity</th>
-            <th>Name</th>
-            <th>{"  "}</th>
-          </tr>
-        </thead>
-        <tbody>
-          { _.map(deckCards, (card) => {
-              return (
-                <>
-                  <tr>
-                    <td>
-                      {card.quantity}
-                    </td>
-                    <td>
-                      <Link
-                        className="link-primary"
-                        data-tip
-                        data-for={`card-${card.id}`}
-                        to={`/cards/${card.id}`}
-                      >
-                        {card.name}
-                      </Link>
-                      <ReactTooltip
-                        backgroundColor="white"
-                        border
-                        borderColor="black"
-                        className="card-tooltip"
-                        textColor="black"
-                        id={`card-${card.id}`}
-                        place="right"
-                      >
-                        <CardPanel cardData={ card } />
-                      </ReactTooltip>
-                    </td>
-                  </tr>
-                </>
-              )
-            })
-          }
-        </tbody>
-      </table>
+      <div className="row">
+        <h2>{ deckData.name }</h2>
+        <div className="col-md-6">
+          <DeckCardList deckData={ deckData } />
+        </div>
+        <div className="col-md-6">
+          <div className="container">
+            { isUserDeck() && <Link to={`/editDeck/${deckData.id}`}><button type="button" className="btn btn-primary mx-1">Edit</button></Link> }
+            {/* TODO <Link><button type="button" className="btn btn-secondary mx-1">Clone</button></Link> */}
+          </div>
+          <div className="row">
+            <div className="fw-bold">
+              Description
+            </div>
+            <div>
+              { deckData.description }
+            </div>
+          </div>
+        </div>
+      </div>
     </div>
   );
 };
